irdeto: remove install test bin [1/1]
PD#SWPL-86027

Problem:
remove install test bin

Solution:
remove install test bin

Verify:
S905C3

(cherry picked from commit 954b1f1b59a989feb0076f68d84047c724096b29)

+++ recipes-cas/irdeto/irdeto-cashal-rel.bb
@@ -5,7 +5,7 @@ PV = "git${SRCPV}"
 PR = "r0"
 
 #Only enable it in OpenLinux
-#SRC_URI_append = " ${@bb.utils.contains('DISTRO_FEATURES', 'irdeto', 'git://${AML_GIT_ROOT_OP}/irdeto-cashal-rel.git;protocol=${AML_GIT_ROOT_PROTOCOL};branch=projects/openlinux/v3.0-rdk','', d)}"
+#SRC_URI_append = " ${@bb.utils.contains('DISTRO_FEATURES', 'irdeto', 'git://${AML_GIT_ROOT_OP}/irdeto-cashal-rel.git;protocol=${AML_GIT_ROOT_PROTOCOL};branch=projects/openlinux/source','', d)}"
 SRC_URI_append = " ${@get_patch_list_with_path('${COREBASE}/aml-patches/../vendor/irdeto/irdeto-cashal-rel')}"
 
 DEPENDS = "liblog aml-secmem aml-mediahal-sdk irdeto-sdk aml-cas-hal cjson optee-userspace"
@@ -19,17 +19,20 @@ LIC_FILES_CHKSUM = "file://${COREBASE}/../meta-meson/license/AMLOGIC;md5=6c70138
 EXTRA_OEMAKE="TARGET_DIR=${D} STAGING_DIR=${STAGING_DIR_HOST}"
 
 do_compile () {
-	oe_runmake  -C ${S} all
+    oe_runmake  -C ${S} all
 }
 
 do_install() {
-    install -d -m 0644 ${D}/${libdir}
-	install -d -m 0644 ${D}/${bindir}
-	install -D -m 0644 ${S}/libird_dvb.so ${D}/${libdir}
-	install -D -m 0644 ${S}/ird_test ${D}/${bindir}
 
-	install -d -m 0644 ${D}/etc/cas/irdeto/cadata
-	install -D -m 0644 ${S}/data/cloaked_ca_*.dat ${D}/etc/cas/irdeto/cadata
+    if [ -e ${S}/libird_dvb.so ] ; then
+        install -d -m 0644 ${D}/${libdir}
+        install -D -m 0644 ${S}/libird_dvb.so ${D}/${libdir}
+    fi
+
+    if [ -e ${S}/data/cloaked_ca_*.dat ] ; then
+        install -d -m 0644 ${D}/etc/cas/irdeto/cadata
+        install -D -m 0644 ${S}/data/cloaked_ca_*.dat ${D}/etc/cas/irdeto/cadata
+    fi
 }
 
 FILES_${PN} = "${libdir}/* /usr/lib/* ${bindir}/* /etc/cas/irdeto/cadata/*"
